requirements.txt
numpy
grpcio
protobuf
grpcio-tools
opencv-python
asyncio
aiohttp
